moved Colors to utils
